common: Add logger API
